fixing o2 c++ error

+++ PWGCF/EbyEFluctuations/Tasks/radialFlowDecorr.cxx
@@ -231,7 +231,7 @@ struct RadialFlowDecorr {
   const AxisSpec etaAxis{{-0.9, -0.8, -0.7, -0.6, -0.5, -0.4, -0.3, -0.2, -0.1, 0.0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9}, "Eta"};
   const AxisSpec phiAxis{KNbinsPhi, KPhiMin, TwoPI, "#phi"};
   const AxisSpec etaBinAxis{KNEta + 1, -0.5, KNEta + 0.5, "#eta bin Number"};
-  const AxisSpec spBinAxis{KNsp + 1, -KBinOffset, KNsp + KBinOffset, "species index Number"};
+  const AxisSpec spBinAxis{KNsp + 1, -KBinOffset, static_cast<float>(KNsp) + KBinOffset, "species index Number"};
 
   const AxisSpec gapAxis{{-1.55, -1.45, -1.35, -1.25, -1.15, -1.05, -0.95, -0.85,
                           -0.75, -0.65, -0.55, -0.45, -0.35, -0.25, -0.15, -0.05,
